Reduce duplication in object store integration tests.
